Add test to ensure a bad incremental strategy results in an error on run
Right now, this test will fail. That is because the fix will be in the
next commit. Presently if an incremental model is being built for the
first time (because the relation in the data warehouse doesn't exist)
or if the incremental model is being run as a full refresh, then validation
that the incremental strategy _doesn't happen_. Technically for _most_
incremental strategies, this is "okay" because the incremental strategy
isn't used when creating the relation for the first time or during full
refresh. The exception is the microbatch strategy, which is part of the
problem. The other part of the problem is that this leads to an experience
where if you write up an incremental model (with an invalid strategy) and
then run it to ensure it works, you won't know anything is wrong until it
is run for the second time.