Raise error for microbatch incremental models in dbt-athena if `partition_by` is `None`
A `partition_by` must be set for the microbatch incremental models in dbt-athena.
Because of this, we will raise an error if `partition_by` is `None` for microbatch
incremental models. This is _different_ from how we handle the same situation for
the `insert_overwrite` strategy in dbt-athena. Honestly, the `insert_overwrite` strategy
should probably do the same thing. However, change it now would possibly break people
and so we won't alter `insert_overwrite` (for the time being).