Complete fix.
Users who wait until Monday (or later by exception) get a latency penalty but that should be excessively few since Snowflake requires users upgrade with no way to change that. This is a complex in interface yet straightforward in code approach to deal with the possible mix of metadata queries in the dbt-snowflake userspace. In time all these hacks and workarounds and all traces of the behavior flag will be removed for good.