Correct comment delineation in dbt-athena's timestamps macro file
The comments were using `{% .. %}` instead of `{# .. #}`. The former
of which is interpreted as containing jinja code. This then was raising
a deprecation warning about an unrecognized jinja block. For people raising
warnings to errors, this deprecation warning was then becoming a parse time
error. Moving to the proper `{# .. #}` comment syntax resolved this issue.