🩹 Remove table_owner from text only cols
PR #1057 added several new column names to the `text_only_columns` arg
to the `table_from_rows` call in the `_catalog_filter_table` class
method. This method is called as part of the `dbt docs generate` task.

However, some of the new column names included are not included in
the list of columns returned by certain adapter implementations for the
`__get_catalog` and `__get_catalog_relations`.
For example, the relevant BigQuery and Athena macros do not return a
column named `table_owner` (see #1135).

This results in the agate type checker emitting a warning when
attempting typing checking runs.

Any columns added to the `text_only_columns` argument for the call to
`table_from_rows` in the base implementation should presumably be
columns that are returned by all adapters implementations.