Future proof volatility macros to catch any future unknown volatilities
Technically it should be impossible for a new volatility to be created
without our knowledge of it and for it to reach dbt-adapters macros. That
is because `volatility` is defined as an `Enum` in core with only three
possible values `deterministic`, `stable`, and `non-deterministic`. But
if a new volatility is invented, and we start supporting setting it in core
then we will want to handle it in dbt-adapters (by default as a warning) until
we add _specific_ handling in dbt-adapters to support it.