Move dbt-redshift's dbt-postgres req spec to `>=1.10.0rc1,<2.0`
dbt-redshift was previously excluding `dbt-postgres>=1.10`. This was a problem
when trying to install pre-releases of dbt-redshift and dbt-postgres in the
same environment as the latest dbt-postgres wasn't viable for dbt-redshift.
Additionally, new versions of redshift will depend on some functionalities in
new versions of dbt-postgres. As such we've updated the spec to
`dbt-postgres>=1.10.0rc1,<2.0`. Before we release the next GA of dbt-redshift,
the `rc1` portion of the req spec should be dropped.